ensure that UTF32 is 32-bits on a 64-bit platform. It should still be 32 on a 32-bit platform even with this change.

+++ ConvertUTF.h
@@ -87,7 +87,7 @@
     bit mask & shift operations.
 ------------------------------------------------------------------------ */
 
-typedef unsigned long	UTF32;	/* at least 32 bits */
+typedef unsigned int	UTF32;	/* at least 32 bits */
 typedef unsigned short	UTF16;	/* at least 16 bits */
 typedef unsigned char	UTF8;	/* typically 8 bits */
 typedef unsigned char	Boolean; /* 0 or 1 */
